webapi/tct-canvas-html5-tests/canvas-py/canvasgradient_addColorStop_INDEX_SIZE_ERR.html

<!DOCTYPE html>
<!--
Copyright (c) 2014 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this lis
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work withou
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
         Cao, Jun <junx.cao@intel.com>

-->

<html>
  <head>
    <title>Canvas Test: canvasgradient_addColorStop_INDEX_SIZE_ERR</title>
    <link rel="author" title="Intel" href="http://www.intel.com/" />
    <link rel="help" href="http://www.w3.org/TR/2012/WD-2dcontext-20120329/#canvasgradient" />
    <meta name="assert" content="Check if the offset is less than 0, greater than 1, infinite, or NaN, then an INDEX_SIZE_ERR exception must be raised"/>
    <script src="../resources/testharness.js"></script>
    <script src="../resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <canvas id="canvas">
      Your browser does not support the canvas element.
    </canvas>
    <script>
        function invalid_args (offset) {
            test(function () {
                var canvas = document.getElementById("canvas");
                var ctx = canvas.getContext("2d");
                var lingrad = ctx.createLinearGradient(0, 0, 100, 0);
                try {
                    lingrad.addColorStop(offset, "#000000");
                    assert_true(false, "not throw an INDEX_SIZE_ERR exception");
                } catch (e) {
                    assert_equals(e.code, DOMException.INDEX_SIZE_ERR, "throw an INDEX_SIZE_ERR exception");
                }
           });
       }
       invalid_args(-1);
       invalid_args(2);
       invalid_args(Infinity);
       invalid_args(NaN);
    </script>
  </body>
</html>
